Fix declaration of uint_native
Its width should be equal to machine word length, which is 64bit on
riscv64.

Fixes: e9fe7de ("Add a dirty little riscv64 demo")

+++ include/efidef.h
@@ -14,7 +14,7 @@ typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long int uint64_t;
-typedef uint32_t uint_native;
+typedef uint64_t uint_native;
 typedef uint8_t bool;
 typedef uint16_t wchar_t;
 typedef unsigned long int size_t;
